requirements.txt
openai==1.65.5
websockets==13.1
loguru==0.7.3
python-dotenv==1.0.1
requests==2.32.3

# Milvus向量数据库相关依赖
pymilvus==2.3.6
sentence-transformers==4.1.0
langchain-openai==0.1.2
# LLM相关
langchain>=0.0.267
llama-index>=0.8.0
llama-index-vector-stores-milvus>=0.1.0


numpy>=1.21.0
pandas>=1.3.0
tqdm>=4.62.0

# 文本处理
jieba>=0.42.1
regex>=2021.8.3

# BM25关键词搜索
rank_bm25>=0.2.2
tabulate>=0.9.0

# 数据处理
scikit-learn>=1.0.0

# 可选：更好的模型支持
torch>=1.9.0
transformers>=4.21.0

# 开发依赖（可选）
jupyter>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0